requirements.txt
aiofiles==23.2.1
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
anyio==4.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.115.3
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python-headless==4.10.0.84
orjson==3.10.10
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.16
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
ruff==0.7.1
scikit-image==0.21.0
scipy==1.10.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.3
tifffile==2023.7.10
tomlkit==0.12.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
zipp==3.20.2
